requirements.txt
requests==2.32.3
boto3==1.37.5
PyJWT==2.10.1